move stf includes again
