+++ code/src/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_library(stf INTERFACE)
 
-target_include_directories(stf INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/include)
+target_include_directories(stf INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/include/interface)
 
 # TODO (stouff) add the public headers inside the include directory to our install
