+++ code/src/include/interface/stf/alg/circle.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "../math/constants.hpp"
-#include "../math/vector.hpp"
+#include "stf/math/constants.hpp"
+#include "stf/math/vector.hpp"
 
 namespace stf {
 namespace alg {
