+++ code/src/include/interface/stf/alg/interpolation.hpp
@@ -4,7 +4,7 @@
 
 #include <algorithm>
 
-#include "../math/vector.hpp"
+#include "stf/math/vector.hpp"
 
 namespace stf {
 namespace alg {
